chore: textarea type 분리
