+++ src/shared/components/Textarea.tsx
@@ -3,7 +3,7 @@ import { Label } from "@/shared/components/Label";
 import DangerIcon from "@/assets/icon/system-danger.svg";
 import CompletedIcon from "@/assets/icon/system-success.svg";
 import FocusIcon from "@/assets/icon/system-info.svg";
-import type { TextAreaProps } from "@/shared/types/textInput.types";
+import type { TextAreaProps } from "@/shared/types/textarea.types";
 
 export const TextArea = forwardRef<HTMLTextAreaElement, TextAreaProps>(
   (
