+++ src/shared/types/textInput.types.ts
@@ -21,13 +21,3 @@ export type LabelProps<E extends React.ElementType> = {
   children: React.ReactNode;
   className?: string;
 } & React.ComponentPropsWithoutRef<E>;
-
-export type TextAreaProps = {
-  id: string;
-  title?: string;
-  description?: string;
-  maxLength?: number;
-  error?: string;
-  completed?: string;
-  focusMessage?: string;
-} & React.TextareaHTMLAttributes<HTMLTextAreaElement>;
