+++ src/shared/types/textarea.types.ts
@@ -1,0 +1,9 @@
+export type TextAreaProps = {
+  id: string;
+  title?: string;
+  description?: string;
+  maxLength?: number;
+  error?: string;
+  completed?: string;
+  focusMessage?: string;
+} & React.TextareaHTMLAttributes<HTMLTextAreaElement>;
